Move db implementation to core.
Use lib directory only to provide API to plugin.
