+++ core/db/document.ts
