+++ core/db/schema/type.ts
@@ -1,0 +1,12 @@
+enum Type {
+  boolean,
+  integer,
+  float,
+  string,
+  date,
+  array,
+  embedding,
+  objectId
+}
+
+export = Type;
