+++ test/db/connection.ts
@@ -1,14 +1,14 @@
 import assert = require('assert');
-import connection = require('../../lib/db/connection');
+import db = require('../../core/db');
 
 describe('db', () => {
   describe('#initialize', () => {
     it('initialize db with url', (done: MochaDone) => {
-      connection.initialize('mongodb://localhost:27017/beyondTest')
+      db.initialize('mongodb://localhost:27017/beyondTest')
       .onComplete((err: Error) => {
         assert.ifError(err);
 
-        connection.close(true).onComplete((err: Error) => {
+        db.close(true).onComplete((err: Error) => {
           assert.ifError(err);
           done();
         });
@@ -16,7 +16,7 @@ describe('db', () => {
     });
 
     it('Failed if url is wrong', (done: MochaDone) => {
-      connection.initialize('mongodb://localhost:2701/beyondTest')
+      db.initialize('mongodb://localhost:2701/beyondTest')
       .onSuccess(() => {
         done(new Error('initialize must fail with wrong url'));
       }).onFailure((err: Error) => {
@@ -28,7 +28,7 @@ describe('db', () => {
 
   describe('#close', () => {
     it('close before initialize', (done: MochaDone) => {
-      connection.close(true).onSuccess(() => {
+      db.close(true).onSuccess(() => {
         done();
       }).onFailure((err: Error) => {
         done(err);
@@ -38,16 +38,16 @@ describe('db', () => {
 
   describe('#connection', () => {
     it('connection is undefined if not initialized', () => {
-      assert.equal(connection.connection(), undefined);
+      assert.equal(db.connection(), undefined);
     });
 
     it('connection is not undefined if initialized', (done: MochaDone) => {
-      connection.initialize('mongodb://localhost:27017/beyondTest')
+      db.initialize('mongodb://localhost:27017/beyondTest')
       .onComplete((err: Error) => {
         assert.ifError(err);
-        assert(connection.connection());
+        assert(db.connection());
 
-        connection.close(true).onComplete((err: Error) => {
+        db.close(true).onComplete((err: Error) => {
           assert.ifError(err);
           done();
         });
